remove javascript package
